Refactor: Extracted OasTypeParser detection rules to separate object
